Moving object grid reducer
