+++ src/modules/control/ControlViewContainer.js
@@ -21,7 +21,7 @@ import { I18n } from 'react-i18nify';
 import { connect } from 'react-redux';
 import { loadControl } from '../../reducers/control';
 import ControlViewComponent from './ControlViewComponent';
-import { loadRoot } from '../storagefacility/reducers/grid';
+import { loadRoot } from '../storagefacility/reducers/grid/nodes';
 
 const mapStateToProps = (state) => ({
   translate: (key, markdown) => I18n.t(key, markdown),
