+++ src/modules/storagefacility/EditNodeContainer.js
@@ -5,7 +5,7 @@ import EditStorageUnitContainer from './EditNodeComponent';
 import { update as updateState } from './reducers/panel/state';
 import { emitError, emitSuccess } from '../../util/errors/emitter';
 import { I18n } from 'react-i18nify';
-import { loadRoot } from './reducers/grid';
+import { loadRoot } from './reducers/grid/nodes';
 
 const mapStateToProps = (state) => {
   return {
