+++ src/reducers/index.js
@@ -23,10 +23,10 @@ import auth from './auth';
 import picklistReducer from './picklist';
 import storagePanelStateReducer from '../modules/storagefacility/reducers/panel/state';
 import storagePanelReducer from '../modules/storagefacility/reducers/panel';
-import storageNodeGridReducer from '../modules/storagefacility/reducers/grid';
+import storageNodeGridReducer from '../modules/storagefacility/reducers/grid/nodes';
 import storageUnitModal from '../modules/storagefacility/reducers/modal';
 import storageUnitStatsReducer from '../modules/storagefacility/reducers/stats';
-import storageObjectGridReducer from './storageobject/grid';
+import storageObjectGridReducer from '../modules/storagefacility/reducers/grid/objects';
 import observationReducer from './observation';
 import reportReducer from './reports';
 import controlReducer from './control';
